fix(sendy): lint issues
