+++ packages/pieces/sendy/src/lib/actions/delete-subscriber.ts
@@ -1,17 +1,17 @@
 import { createAction, Property, Validators } from "@activepieces/pieces-framework";
-import { status } from "../api";
+import { deleteSubscriber } from "../api";
 import { buildListDropdown } from "../props";
 import { sendyAuth, SendyAuthType } from "../auth";
 
-export const statusAction = createAction({
-	name        : 'status',
+export const deleteAction = createAction({
+	name        : 'delete_subscriber',
 	auth        : sendyAuth,
-	displayName : 'Subscription Status',
-	description : 'Get the subscription status of a user',
+	displayName : 'Delete Subscriber',
+	description : 'Delete a subscriber from a list',
 	props       : {
 		list: Property.Dropdown({
 			displayName : 'List',
-			description : 'Select the list to get the status from',
+			description : 'Select the list to delete from',
 			required    : true,
 			refreshers  : ['auth'],
 			options     : async ({auth}) => await buildListDropdown(auth as SendyAuthType),
@@ -24,7 +24,7 @@ export const statusAction = createAction({
 		}),
 	},
 	async run(context) {
-		return await status(context.auth, {
+		return await deleteSubscriber(context.auth, {
 			list_id : context.propsValue.list,
 			email   : context.propsValue.email,
 		});
