+++ packages/pieces/sendy/src/lib/actions/get-brand-lists.ts
@@ -3,18 +3,11 @@ import { getLists } from "../api";
 import { sendyAuth } from "../auth";
 
 export const getListsAction = createAction({
-	name        : 'getLists',
+	name        : 'get_brand_lists',
 	auth        : sendyAuth,
-	displayName : 'Get Lists',
+	displayName : 'Get Lists for a Brand',
 	description : 'Get the Lists for a Brand',
 	props       : {
-		// brandId: Property.Dropdown({
-		// 	displayName : 'Brand',
-		// 	description : 'Select the brand to get lists for',
-		// 	required    : true,
-		// 	refreshers  : ['auth'],
-		// 	options     : async ({auth}) => await buildBrandDropdown(auth as SendyAuthType),
-		// }),
 		includeHidden: Property.Checkbox({
 			displayName  : 'Include Hidden Lists',
 			description  : 'Include hidden lists in the results',
