+++ packages/pieces/sendy/src/lib/actions/get-brands.ts
@@ -3,7 +3,7 @@ import { getBrands } from "../api";
 import { sendyAuth } from "../auth";
 
 export const getBrandsAction = createAction({
-	name        : 'getBrands',
+	name        : 'get_brands',
 	auth        : sendyAuth,
 	displayName : 'Get Brands',
 	description : 'Get a list of brands from Sendy',
