+++ packages/pieces/sendy/src/lib/actions/get-subscription-status.ts
@@ -1,17 +1,17 @@
 import { createAction, Property, Validators } from "@activepieces/pieces-framework";
-import { deleteSubscriber } from "../api";
+import { status } from "../api";
 import { buildListDropdown } from "../props";
 import { sendyAuth, SendyAuthType } from "../auth";
 
-export const deleteAction = createAction({
-	name        : 'deleteSubscriber',
+export const statusAction = createAction({
+	name        : 'get_subscription_status',
 	auth        : sendyAuth,
-	displayName : 'Delete Subscriber',
-	description : 'Delete a subscriber from a list',
+	displayName : 'Get Subscription Status',
+	description : 'Get the subscription status of a user',
 	props       : {
 		list: Property.Dropdown({
 			displayName : 'List',
-			description : 'Select the list to delete from',
+			description : 'Select the list to get the status from',
 			required    : true,
 			refreshers  : ['auth'],
 			options     : async ({auth}) => await buildListDropdown(auth as SendyAuthType),
@@ -24,7 +24,7 @@ export const deleteAction = createAction({
 		}),
 	},
 	async run(context) {
-		return await deleteSubscriber(context.auth, {
+		return await status(context.auth, {
 			list_id : context.propsValue.list,
 			email   : context.propsValue.email,
 		});
